requirements.txt
asn1crypto==0.24.0
certifi==2019.6.16
cffi==1.12.3
colorama==0.4.1
cryptography==2.7
emoji==0.5.3
future==0.17.1
pycparser==2.19
python-dotenv==0.10.3
python-telegram-bot==12.0.0
six==1.12.0
tornado==6.0.3
